add tutorial notebook and fixup ingresses
