+++ buildingmotif/ingresses/__init__.py
@@ -1,9 +1,8 @@
 from buildingmotif.ingresses.bacnet import BACnetNetwork  # noqa
 from buildingmotif.ingresses.base import Record  # noqa
 from buildingmotif.ingresses.brick import BACnetToBrickIngress  # noqa
-from buildingmotif.ingresses.csv import (  # noqa
-    CSVIngress,
+from buildingmotif.ingresses.csv import CSVIngress, XLSXIngress  # noqa
+from buildingmotif.ingresses.templates import (  # noqa
     TemplateIngress,
     TemplateIngressWithChooser,
-    XLSXIngress,
 )
